client doesn't crash immediately!
when transitioning from single player to multiplayer, some
assumptions you made for single player will not make sense, and
sometimes cause the program to blow up.  In this case, the assumption
that when the GUI is being shown, the players have all joined
(and therefore the button can query the humanPlayer) is no longer valid.
When fixing these bugs, make sure you don't break it for single player.